std err redirection test
